+++ returns/maybe.py
@@ -1,17 +1,17 @@
 # -*- coding: utf-8 -*-
 
 from abc import ABCMeta
-from typing import Generic, TypeVar
+from typing import TypeVar
 
 from typing_extensions import Literal
 
 from returns.primitives.exceptions import UnwrapFailedError
-from returns.primitives.monad import Monad
+from returns.primitives.monad import GenericMonadOneSlot
 
 _ValueType = TypeVar('_ValueType')
 
 
-class Maybe(Generic[_ValueType], Monad[_ValueType], metaclass=ABCMeta):
+class Maybe(GenericMonadOneSlot[_ValueType], metaclass=ABCMeta):
     """
     Represents a result of a series of commutation that can return ``None``.